// src/schemas/RemoteApplicationToggleRequest.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Remote application toggle request */
export type RemoteApplicationToggleRequest = {
  /** Standard ROS header */
  header: Header;

  app_toggle_req_speed_limit4kmh: boolean;
  app_toggle_req_send_c_a_n_dbg_messages: boolean;
  app_toggle_req_pwr_assisted_braking: boolean;
};