// src/schemas/RemoteIndicatorRequest.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Remote indicator request */
export type RemoteIndicatorRequest = {
  /** Standard ROS header */
  header: Header;

  remote_status_light_blink_rate_req: number;
  remote_status_light_color_req: number;
  remote_horn_req: boolean;
  remote_light_brake_req: boolean;
  remote_blink_left_req: boolean;
  remote_blink_right_req: boolean;
  remote_brights_on_req: boolean;
  remote_light_reverse_req: boolean;
  remote_lights_turn_left_req: boolean;
  remote_lights_turn_right_req: boolean;
};